fix: rename value can not be chinese

+++ src/utils/config-manager.ts
@@ -2,15 +2,21 @@ import { basename, join } from "@tauri-apps/api/path";
 import { MediaEditorSchema } from "../schemas/project-config";
 import { exists, mkdir, readDir, readTextFile, rename, writeTextFile } from "@tauri-apps/plugin-fs";
 import { getVersion } from "@tauri-apps/api/app";
-import { configFileName, getUniquePath, keyFileName, projectConfigDirPath } from "./path";
-import { getCurrentDate } from "./common";
-import { MediaEditorProject } from "../types/project-config";
+import { configFileName, getUniquePath, keyFileName, appProjectConfigDirPath } from "./path";
+import { copyText, getCurrentDate } from "./common";
+import { MediaEditorConfig, MediaEditorProject } from "../types/project-config";
 import { invoke } from "@tauri-apps/api/core";
 import { generateKey, obfuscateData, deobfuscateData } from "./cypher";
 
-export async function persistProjectConfig(config: MediaEditorProject, projectDirPath: string) {
+export async function persistProjectConfig(project: MediaEditorProject) {
+    const { config, name } = project
+    const projectDirPath = await join(appProjectConfigDirPath, name)
     const key = generateKey(16);
     const obfuscatedConfig = obfuscateData(config, key);
+
+    if (!await exists(projectDirPath)) {
+        throw new Error(`project path: "${projectDirPath}" is not exists`)
+    }
 
     // write the config file
     await writeTextFile(await join(projectDirPath, configFileName), obfuscatedConfig);
@@ -30,15 +36,14 @@ export async function loadProjectConfig(projectDirPath: string) {
     }
 }
 
-export async function genertateEditorConfig(name: string) {
+export async function genertateEditorConfig() {
     const now = new Date().toLocaleString();
 
-    const config: MediaEditorProject = {
+    const config: MediaEditorConfig = {
         assets: [],
         metadata: {
             created_at: now,
             last_modified: now,
-            name: name,
             version: await getVersion(),
             cover_path: null
         },
@@ -50,39 +55,45 @@ export async function genertateEditorConfig(name: string) {
     return MediaEditorSchema.parse(config);
 }
 
-async function createProjectDir(name: string) {
-    const uniquePath = await getUniquePath(projectConfigDirPath, name);
+async function createUniqueProjectDir(name: string) {
+    const uniquePath = await getUniquePath(appProjectConfigDirPath, name);
     await mkdir(uniquePath, { recursive: true });
     return uniquePath
 }
 
 export async function copyProject(project: MediaEditorProject) {
-    const projectPath = await createProjectDir(project.metadata.name)
+    const projectPath = await createUniqueProjectDir(`${project.name}-${copyText()}`)
     const now = new Date().toLocaleString();
     const newMetadata = {
-        ...project.metadata,
+        ...project.config.metadata,
         name: await basename(projectPath),
         last_modified: now,
         created_at: now,
         version: await getVersion(),
     }
-    await persistProjectConfig({ ...project, metadata: newMetadata }, projectPath)
-    return loadProjectConfig(projectPath)
+
+    await persistProjectConfig({ name: project.name, config: { ...project.config, metadata: newMetadata } })
+
+    const config = await loadProjectConfig(projectPath)
+    if (!config) {
+        throw new Error("can not load copyed project")
+    }
+    return { name: await basename(projectPath), config }
 }
 
 export async function createNewProject() {
-    // ceate project directory    
-    const uniquePath = await createProjectDir(getCurrentDate())
+    // create project directory    
+    const uniquePath = await createUniqueProjectDir(getCurrentDate())
+    const uniqueName = await basename(uniquePath)
 
-    // generate project config file
-    const projectName = await basename(uniquePath);
-    const projectConfig = await genertateEditorConfig(projectName);
-    await persistProjectConfig(projectConfig, uniquePath);
+    // generate project config file    
+    const projectConfig = await genertateEditorConfig();
+    await persistProjectConfig({ config: projectConfig, name: uniqueName });
     return projectConfig
 }
 
 export async function deleteProject(name: string) {
-    const projectPath = await join(projectConfigDirPath, name)
+    const projectPath = await join(appProjectConfigDirPath, name)
     if (!await exists(projectPath)) {
         return;
     }
@@ -95,25 +106,33 @@ export async function deleteProject(name: string) {
 }
 
 export async function listAllProjects() {
-    const entryList = await readDir(projectConfigDirPath);
-    return (await Promise.all(entryList.filter(entry => entry.isDirectory)
-        .map(async (entry) =>
-            await loadProjectConfig(await join(projectConfigDirPath, entry.name))
-        )
-    )).filter(project => project !== null);
+    const entryList = await readDir(appProjectConfigDirPath);
+    const dirEntryList = entryList.filter(entry => entry.isDirectory)
+    const projectResultList = await Promise.all(
+        dirEntryList.map(async (entry) => ({
+            name: entry.name,
+            config: await loadProjectConfig(await join(appProjectConfigDirPath, entry.name))
+        }))
+    )
+    return projectResultList.filter((project): project is { name: string, config: MediaEditorConfig } => project.config !== null)
 }
 
-export async function updateProject(project: MediaEditorProject, oldName: string) {
-    const oldProjectDirPath = await join(projectConfigDirPath, oldName)
+export async function updateProject(project: MediaEditorProject) {
+    await persistProjectConfig(project)
+}
+
+export async function renameProject(oldName: string, newName: string) {
+    const oldProjectDirPath = await join(appProjectConfigDirPath, oldName)
+
     if (!await exists(oldProjectDirPath)) {
         throw new Error("the old project config path is not exsits")
     }
 
-    const newProjectDirPath = await join(projectConfigDirPath, project.metadata.name);
+    const newProjectDirPath = await join(appProjectConfigDirPath, newName)
+
     if (await exists(newProjectDirPath)) {
         throw new Error("the new project config path is already exists")
     }
 
     await rename(oldProjectDirPath, newProjectDirPath)
-    await persistProjectConfig(project, newProjectDirPath)
 }
